requirements.txt
apache-airflow-providers-amazon==7.1.0